requirements.txt
beautifulsoup4
requests
pydash
slackclient